Use "name-value" for path segment parameters
