+++ library/CM/Url/ResourceUrl.php
@@ -2,9 +2,10 @@
 
 namespace CM\Url;
 
+use CM_Site_Abstract;
 use CM_Frontend_Environment;
 
-class ResourceUrl extends AssetUrl {
+class ResourceUrl extends Url {
 
     /** @var string */
     protected $_type;
@@ -23,21 +24,16 @@ class ResourceUrl extends AssetUrl {
         $this->_type = (string) $type;
     }
 
-    public function getUriRelativeComponents() {
-        $segments = [
-            $this->getType(),
-        ];
-        if ($language = $this->getLanguage()) {
-            $segments[] = $language->getAbbreviation();
-        }
-        if ($site = $this->getSite()) {
-            $segments[] = $site->getId();
-        }
-        if ($deployVersion = $this->getDeployVersion()) {
-            $segments[] = $deployVersion;
-        }
-        $segments = array_merge($segments, $this->_getPathSegments());
-        return $this->_getPathFromSegments($segments) . $this->_getQueryComponent() . $this->_getFragmentComponent();
+    public function withSite(CM_Site_Abstract $site) {
+        $url = clone $this;
+        $url->_site = $site;
+        return $url->withBaseUrl($site->getUrlCdn());
+    }
+
+    protected function _getParameterSegments() {
+        $segments = parent::_getParameterSegments();
+        $segments[] = $this->getType();
+        return $segments;
     }
 
     /**
@@ -49,7 +45,7 @@ class ResourceUrl extends AssetUrl {
      */
     public static function create($url, $type, CM_Frontend_Environment $environment = null, $deployVersion = null) {
         /** @var ResourceUrl $resourceUrl */
-        $resourceUrl = parent::_create($url, $environment, $deployVersion);
+        $resourceUrl = parent::createWithEnvironment($url, $environment, $deployVersion);
         $resourceUrl->setType($type);
         return $resourceUrl;
     }
