+++ library/CM/Url/RouteUrl.php
@@ -4,18 +4,7 @@ namespace CM\Url;
 
 use CM_Frontend_Environment;
 
-class RouteUrl extends AbstractUrl {
-
-    public function getUriRelativeComponents() {
-        $segments = $this->_getPathSegments();
-        if ($prefix = $this->getPrefix()) {
-            $segments = array_merge([$prefix], $segments);
-        }
-        if ($language = $this->getLanguage()) {
-            $segments[] = $language->getAbbreviation();
-        }
-        return $this->_getPathFromSegments($segments) . $this->_getQueryComponent() . $this->_getFragmentComponent();
-    }
+class RouteUrl extends Url {
 
     /**
      * @param string                       $route
@@ -25,7 +14,7 @@ class RouteUrl extends AbstractUrl {
      */
     public static function create($route, array $params = null, CM_Frontend_Environment $environment = null) {
         /** @var RouteUrl $url */
-        $url = parent::_create($route, $environment);
+        $url = parent::createWithEnvironment($route, $environment);
         if (null !== $params) {
             $url = $url->withParams($params);
         }
