+++ tests/library/CM/Url/ResourceUrlTest.php
@@ -13,15 +13,15 @@ class ResourceUrlTest extends CMTest_TestCase {
     }
 
     public function testCreate() {
-        $url = ResourceUrl::create('file.ext', 'resource-type');
-        $this->assertSame('/resource-type/file.ext', (string) $url);
+        $url = ResourceUrl::create('file.ext', 'resource/type');
+        $this->assertSame('/resource/type/file.ext', (string) $url);
 
         $environment = $this->createEnvironment(['type' => 999], null, 'de');
-        $url = ResourceUrl::create('file.ext', 'resource-type', $environment);
-        $this->assertSame('http://cdn.example.com/resource-type/de/999/file.ext', (string) $url);
+        $url = ResourceUrl::create('file.ext', 'resource/type', $environment);
+        $this->assertSame('http://cdn.example.com/language-de/site-999/resource/type/file.ext', (string) $url);
 
-        $url = ResourceUrl::create('file.ext', 'resource-type', $environment, 1234);
-        $this->assertSame('http://cdn.example.com/resource-type/de/999/1234/file.ext', (string) $url);
+        $url = ResourceUrl::create('file.ext', 'resource/type', $environment, 1234);
+        $this->assertSame('http://cdn.example.com/language-de/site-999/version-1234/resource/type/file.ext', (string) $url);
     }
 
     public function testWithSite() {
@@ -46,9 +46,9 @@ class ResourceUrlTest extends CMTest_TestCase {
             ->method('getUrlCdn')
             ->will($this->returnValue('http://cdn.foo.com/path?param'));
 
-        $url = ResourceUrl::create('file.ext', 'resource-type');
+        $url = ResourceUrl::create('file.ext', 'resource/type');
 
         $urlWithSite = $url->withSite($site);
-        $this->assertSame('http://cdn.foo.com/resource-type/42/file.ext', (string) $urlWithSite);
+        $this->assertSame('http://cdn.foo.com/path/site-42/resource/type/file.ext', (string) $urlWithSite);
     }
 }
